requirements.txt
python-dotenv>=1.0.0
twilio>=9.0.0      # Optional — only needed for SMS
